#16 - separared menu-bar.fxml and status-log.fxml
